(미완성된)XML로드 개선완료

+++ 7.Data(XML)/OfficeData.cs
@@ -22,8 +22,8 @@ public class OfficeData : MonoBehaviour
         }
     }
 
-    private float officeGage;
-    public float OfficeGage{get{return officeGage;}
+    private int officeGage;
+    public int OfficeGage{get{return officeGage;}
         set{
             officeGage = value;
             if(officeGage < 0) officeGage = 0;
